requirements.txt
adversarial-robustness-toolbox==1.20.1
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.11.0
astroid==3.3.11
attrs==25.3.0
Authlib==1.6.5
bandit==1.8.6
black==25.9.0
boltons==21.0.0
boolean.py==5.0
bracex==2.6
Brotli==1.1.0
CacheControl==0.14.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
click-option-group==0.5.8
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.2
cycler==0.12.1
cyclonedx-python-lib==9.1.0
defusedxml==0.7.1
dill==0.4.0
dparse==0.6.4
eagerpy==0.30.0
exceptiongroup==1.2.2
face==24.0.0
fastapi==0.118.0
ffmpy==0.6.1
filelock==3.19.1
fonttools==4.60.1
foolbox==3.3.4
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
glom==22.1.0
googleapis-common-protos==1.70.0
gradio==5.48.0
gradio_client==1.13.3
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.3
idna==3.10
importlib_metadata==8.7.0
isort==6.1.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.20.0
jsonschema-specifications==2025.9.1
kagglehub==0.3.13
kiwisolver==1.4.9
license-expression==30.4.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.0.1
matplotlib==3.10.6
mccabe==0.7.0
mcp==1.12.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.1
mypy_extensions==1.1.0
networkx==3.5
nltk==3.9.2
numpy==2.3.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-requests==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
orjson==3.11.3
packageurl-python==0.17.5
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
peewee==3.18.2
pillow==11.3.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.4.0
protobuf==6.32.1
psutil==7.1.0
py-serializable==2.1.0
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pylint==3.3.8
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rich==13.5.3
rpds-py==0.27.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
ruff==0.13.3
safehttpx==0.1.6
safety==3.6.2
safety-schemas==0.0.16
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
semantic-version==2.10.0
semgrep==1.139.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sse-starlette==3.0.2
starlette==0.48.0
stevedore==5.5.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.0.2
tomlkit==0.13.3
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
triton==3.4.0
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
wcmatch==8.5.2
websockets==15.0.1
wrapt==1.17.3
zipp==3.23.0
